agent/requirements.txt
# Python dependencies
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
pydantic==2.8.2
requests==2.31.0
solana==0.35.0
twitter-api-client==0.10.22
solders
openai
python-dotenv
numpy
tweepy
eth_keys
web3
base58
